+++ R/single_sample_input_qc.R
@@ -1,5 +1,90 @@
 ## -- Functions related to single-sample analysis / stage 01_input_qc.
 
+#' @title Import scRNA-seq data.
+#' @description For more details see the `INPUT_DATA` parameter in `vignette("stage_input_qc")`.
+#' @param input_data A named list of named lists containing character scalars:
+#'   - `type`:
+#'     - `"cellranger"`: a raw feature-barcode matrix from 10x Genomics `cellranger`.
+#'     - `"table"`: a delimited text file (table).
+#'     - `"sce"`: a `SingleCellExperiment` object (Rds file).
+#'   - `path`: a path to input file (or directory in case of `type = "cellranger"`).
+#'   - `delimiter`: a field delimiter when `type = "table"`.
+#' @return A `SingleCellExperiment` object. *Output target*: `sce_raw`
+#'
+#' @concept single_sample_input_qc_fn
+#' @export
+sce_raw_fn <- function(input_data) {
+  input_type <- input_data$type
+  input_path <- input_data$path
+
+  assert_that_(
+    !is_null(input_type), !is_null(input_path),
+    msg = "{.field INPUT_DATA$type} or {.field INPUT_DATA$path} is not set, cannot load data."
+  )
+
+  if (input_type == "cellranger") {
+    sce_raw <- DropletUtils::read10xCounts(input_path)
+    colnames(sce_raw) <- colData(sce_raw)$Barcode
+  } else if (input_type == "table") {
+    delimiter <- input_data$delimiter
+    assert_that_(!is_null(delimiter), msg = "{.field INPUT_DATA$delimiter} is not set, cannot load {.file {input_path}}")
+    sce_raw <- readr::read_delim(input_path, delim = delimiter)
+    assert_that_(
+      "ENSEMBL" %in% colnames(sce_raw),
+      msg = "Column {.var ENSEMBL} with ENSEMBL IDs is missing in {.file {input_path}}"
+    )
+    sce_raw <- as.data.frame(sce_raw) %>%
+      tibble::column_to_rownames("ENSEMBL") %>%
+      as.matrix()
+    mode(sce_raw) <- "integer"
+    sce_raw <- Matrix::Matrix(sce_raw, sparse = TRUE)
+    sce_raw <- SingleCellExperiment(assays = list(counts = sce_raw))
+    ## -- This will add the same information as when a raw BC matrix from cellranger is loaded with
+    ## -- DropletUtils::read10xCounts(), but we probably don't need to do so.
+    # rowData(sce_raw) <- DataFrame(ID = rownames(sce_raw), Symbol = NA, Type = "Gene Expression")
+    # colData(sce_raw) <- DataFrame(Sample = input_path, Barcode = colnames(sce_raw))
+  } else if (input_type == "sce") {
+    sce_raw <- readRDS(input_path)
+    assert_that_(
+      is(sce_raw, "SingleCellExperiment"),
+      msg = "Object loaded from {.file {input_path}} is not a {.var SingleCellExperiment} object."
+    )
+    assert_that_(
+      "counts" %in% assayNames(sce_raw),
+      msg = "{.field counts} assay not found in SCE object loaded from {.file {input_path}}"
+    )
+  } else {
+    cli::cli_abort("{.var input_data$type} must be {.field 'cellranger'}, {.field 'table'}, or {.field 'sce'}")
+  }
+
+  return(sce_raw)
+}
+
+#' @title Calculate empty droplet statistics for each cell.
+#' @description See `DropletUtils::emptyDrops()` for implementation details.
+#' @param sce_raw (*input target*) A `SingleCellExperiment` object.
+#' @param empty_droplets_lower An integer scalar: lower bound on the total UMI count at or below which all barcodes
+#'   are assumed to correspond to empty droplets. Passed to `DropletUtils::emptyDrops()`.
+#' @param empty_droplets_enabled A logical scalar: if `TRUE`, skip empty droplets calculation and return `NULL`.
+#' @inheritParams bpparam_
+#' @return A `DataFrame` object (value returned from `DropletUtils::emptyDrops()`) if `empty_droplets_enabled`
+#' is `TRUE`, `NULL` otherwise. *Output target*: `empty_droplets`
+#'
+#' @concept single_sample_input_qc_fn
+#' @export
+empty_droplets_fn <- function(sce_raw,
+                              empty_droplets_lower,
+                              empty_droplets_enabled = TRUE,
+                              BPPARAM = BiocParallel::SerialParam()) {
+  if (empty_droplets_enabled) {
+    DropletUtils::emptyDrops(
+      m = counts(sce_raw), lower = empty_droplets_lower, BPPARAM = BPPARAM
+    )
+  } else {
+    NULL
+  }
+}
+
 #' @title Subset cells in a `SingleCellExperiment` object to non-empty ones and add corresponding statistics.
 #' @param sce_raw (*input target*) A `SingleCellExperiment` object.
 #' @param empty_droplets (*input target*) A `DataFrame` object returned from [DropletUtils::emptyDrops()].
@@ -9,31 +94,39 @@
 #' @concept single_sample_input_qc_fn
 #' @export
 sce_valid_cells_fn <- function(sce_raw, empty_droplets, empty_droplets_fdr_threshold) {
-  is_cell_index <- which(empty_droplets$FDR <= empty_droplets_fdr_threshold)
-  sce_valid_cells <- sce_raw[, is_cell_index]
-  sce_valid_cells$is_empty_fdr <- empty_droplets$FDR[is_cell_index]
-  colnames(sce_valid_cells) <- sce_valid_cells$Barcode
+  sce_valid_cells <- sce_raw
+
+  if (is_null(empty_droplets)) {
+    sce_valid_cells$is_empty_fdr <- NA
+  } else {
+    is_cell_index <- which(empty_droplets$FDR <= empty_droplets_fdr_threshold)
+    colnames_orig <- colnames(sce_valid_cells)
+    sce_valid_cells <- sce_valid_cells[, is_cell_index]
+    sce_valid_cells$is_empty_fdr <- empty_droplets$FDR[is_cell_index]
+  }
 
   return(sce_valid_cells)
 }
 
-#' @title Get a logical filter for ratio of cells expressing a gene and minimum number of UMI per gene.
+#' @title Get a logical filter for genes not passing a ratio of cells expressing a gene and a minimum number of UMI per gene.
 #' @description A gene is considered expressed when number of its UMIs across all cells is greater than `min_umi` and
 #' at the same time it is expressed in at least `min_ratio_cells` ratio of cells.
 #' @param sce A `SingleCellExperiment` object.
 #' @param min_ratio_cells A numeric scalar: minimum ratio of cells expressing a gene.
 #' @param min_umi A numeric scalar: minimum number of UMI per gene across all cells.
-#' @return A logical vector.
+#' @return A logical vector of length `nrow(sce)`: `TRUE` for genes **not passing** the filter, `FALSE` otherwise.
+#' This is for consistency with cell filters, e.g. the `qc_filter` target, where `TRUE` means the same as here,
+#' i.e. a cell is not passing the filter.
 #'
 #' @concept single_sample_input_qc_fn
 #' @export
 get_gene_filter <- function(sce, min_ratio_cells, min_umi) {
   assert_that_(dplyr::between(min_ratio_cells, 0, 1))
-  assert_that_(min_umi > 0)
+  assert_that_(min_umi >= 0)
 
   num_cells <- min_ratio_cells * ncol(sce)
   is_expressed <- rowSums(counts(sce) >= min_umi) >= num_cells
-  return(is_expressed)
+  return(!is_expressed)
 }
 
 #' @title Create a tibble with history of cell and gene filtering.